# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |-
    Connect with your friends effortlessly using WASAText! Send and receive messages, 
    whether one-on-one or in groups, all from the convenience of your PC. 
    Enjoy seamless conversations with text or GIFs and easily stay in touch 
    through your private chats or group discussions.
  version: "1.0.0"
...
servers:
  - url: "http://localhost:3000"
    description: Development server
paths:
...
  /conversations/{conversationId}/messages/{messageId}/comments:
    parameters:
      - $ref: '#/components/parameters/conversationId'
      - $ref: '#/components/parameters/messageId'
    post:
      tags: ["Messages"]
      summary: Comment on a message
      description: React to a message with an emoticon comment
      operationId: commentMessage
      security:
        - BearerAuth: []
      requestBody:
        description: Comment content
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for adding emoji reaction to a message
              properties:
                emoticon:
                  type: string
                  description: Emoji character to use as reaction
                  pattern: '^.+$'
                  minLength: 1
                  maxLength: 10
                  example: "😀"
              required:
                - emoticon
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
              example:
                id: "comment123"
                userId: "user456"
                userName: "Marco"
                emoticon: "😀"
                timestamp: "2023-12-07T10:35:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/messages/{messageId}/comments/{commentId}:
    parameters:
      - $ref: '#/components/parameters/conversationId'
      - $ref: '#/components/parameters/messageId'
      - $ref: '#/components/parameters/commentId'
    delete:
      tags: ["Messages"]
      summary: Delete a comment
      description: Remove a comment/reaction from a message
      operationId: uncommentMessage
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Comment:
      type: object
      description: Emoji reaction/comment on a message
      properties:
        id:
          type: string
          description: Unique comment identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "comment123"
        userId:
          type: string
          description: ID of the user who made the comment
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "user456"
        userName:
          type: string
          description: Display name of the user who made the comment
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "John"
        emoticon:
          type: string
          description: Emoji character used as reaction
          pattern: '^.+$'
          minLength: 1
          maxLength: 10
          example: "😀"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the comment was made
          minLength: 1
          maxLength: 50
          example: "2023-12-07T10:35:00Z"
      required:
        - id
        - userId
        - userName
        - emoticon
        - timestamp
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (e.g. missing fields, etc.)
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Unauthorized:
      description: The access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Forbidden:
      description: The user is not authorized to perform this operation
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |-
        Bearer authentication using user identifier.
        Use the identifier returned from the login endpoint.
  parameters:
...
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Conversation identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "conv123"
    messageId:
      name: messageId
      in: path
      required: true
      description: Message identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "msg123"
    commentId:
      name: commentId
      in: path
      required: true
      description: Comment identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "comment123"
...